+++ lib/map/layer/image_layer.dart
@@ -3,12 +3,19 @@ import '../../util/extensions.dart';
 
 class ImageLayer extends MapLayer {
   String image;
+  int parallaxX;
+  int parallaxY;
 
   ImageLayer({
     required this.image,
+    this.parallaxX = 1,
+    this.parallaxY = 1,
   });
 
-  ImageLayer.fromJson(Map<String, dynamic> json) : image = json['image'] {
+  ImageLayer.fromJson(Map<String, dynamic> json)
+      : image = json['image'],
+        parallaxX = int.tryParse(json['parallaxx'].toString()) ?? 1,
+        parallaxY = int.tryParse(json['parallaxy'].toString()) ?? 1 {
     id = json['id'];
     name = json['name'];
     offsetX = double.tryParse(json['offsetx'].toString()) ?? 0.0;